refactor: extract country codes map and add logger
