+++ app/src/lib/countryCodes.ts
@@ -1,0 +1,224 @@
+export const COUNTRY_CODE_MAP: Record<string, string> = {
+  Afghanistan: "af",
+  Albania: "al",
+  Algeria: "dz",
+  Andorra: "ad",
+  Angola: "ao",
+  "Antigua and Barbuda": "ag",
+  Argentina: "ar",
+  Armenia: "am",
+  Australia: "au",
+  Austria: "at",
+  Azerbaijan: "az",
+  Bahamas: "bs",
+  Bahrain: "bh",
+  Bangladesh: "bd",
+  Barbados: "bb",
+  Belarus: "by",
+  Belgium: "be",
+  Belize: "bz",
+  Benin: "bj",
+  Bhutan: "bt",
+  Bolivia: "bo",
+  "Bosnia and Herzegovina": "ba",
+  Botswana: "bw",
+  Brazil: "br",
+  Brunei: "bn",
+  Bulgaria: "bg",
+  "Burkina Faso": "bf",
+  Burundi: "bi",
+  "Cabo Verde": "cv",
+  Cambodia: "kh",
+  Cameroon: "cm",
+  Canada: "ca",
+  "Central African Republic": "cf",
+  Chad: "td",
+  Chile: "cl",
+  China: "cn",
+  Colombia: "co",
+  Comoros: "km",
+  Congo: "cg",
+  "Costa Rica": "cr",
+  Croatia: "hr",
+  Cuba: "cu",
+  Curacao: "cw",
+  Cyprus: "cy",
+  Czechia: "cz",
+  "Czech Republic": "cz",
+  "Christmas Island": "cx",
+  "Democratic Republic of the Congo": "cd",
+  Denmark: "dk",
+  Djibouti: "dj",
+  Dominica: "dm",
+  "Dominican Republic": "do",
+  Ecuador: "ec",
+  Egypt: "eg",
+  "El Salvador": "sv",
+  "Equatorial Guinea": "gq",
+  Eritrea: "er",
+  Estonia: "ee",
+  Eswatini: "sz",
+  Ethiopia: "et",
+  Fiji: "fj",
+  Finland: "fi",
+  France: "fr",
+  Gabon: "ga",
+  Gambia: "gm",
+  Georgia: "ge",
+  Germany: "de",
+  Ghana: "gh",
+  Greece: "gr",
+  Grenada: "gd",
+  Guatemala: "gt",
+  Guinea: "gn",
+  "Guinea-Bissau": "gw",
+  Guyana: "gy",
+  Haiti: "ht",
+  Honduras: "hn",
+  Hungary: "hu",
+  Iceland: "is",
+  India: "in",
+  Indonesia: "id",
+  Iran: "ir",
+  Iraq: "iq",
+  Ireland: "ie",
+  Israel: "il",
+  Italy: "it",
+  Jamaica: "jm",
+  Japan: "jp",
+  Jordan: "jo",
+  Kazakhstan: "kz",
+  Kenya: "ke",
+  Kiribati: "ki",
+  Kuwait: "kw",
+  Kyrgyzstan: "kg",
+  Laos: "la",
+  Latvia: "lv",
+  Lebanon: "lb",
+  Lesotho: "ls",
+  Liberia: "lr",
+  Libya: "ly",
+  Liechtenstein: "li",
+  Lithuania: "lt",
+  Luxembourg: "lu",
+  Madagascar: "mg",
+  Malawi: "mw",
+  Malaysia: "my",
+  Maldives: "mv",
+  Mali: "ml",
+  Malta: "mt",
+  "Marshall Islands": "mh",
+  Mauritania: "mr",
+  Mauritius: "mu",
+  Mexico: "mx",
+  Micronesia: "fm",
+  Moldova: "md",
+  Monaco: "mc",
+  Mongolia: "mn",
+  Montenegro: "me",
+  Morocco: "ma",
+  Mozambique: "mz",
+  Myanmar: "mm",
+  Namibia: "na",
+  Nauru: "nr",
+  Nepal: "np",
+  Netherlands: "nl",
+  "New Zealand": "nz",
+  Nicaragua: "ni",
+  Niger: "ne",
+  Nigeria: "ng",
+  "North Korea": "kp",
+  "North Macedonia": "mk",
+  Norway: "no",
+  Oman: "om",
+  Pakistan: "pk",
+  Palau: "pw",
+  "Palestine State": "ps",
+  Panama: "pa",
+  "Papua New Guinea": "pg",
+  Paraguay: "py",
+  Peru: "pe",
+  Philippines: "ph",
+  Poland: "pl",
+  Portugal: "pt",
+  "Puerto Rico": "pr",
+  Qatar: "qa",
+  Romania: "ro",
+  Russia: "ru",
+  Rwanda: "rw",
+  "Saint Kitts and Nevis": "kn",
+  "Saint Lucia": "lc",
+  "Saint Vincent and the Grenadines": "vc",
+  Samoa: "ws",
+  "San Marino": "sm",
+  "Sao Tome and Principe": "st",
+  "Saudi Arabia": "sa",
+  Senegal: "sn",
+  Serbia: "rs",
+  Seychelles: "sc",
+  "Sierra Leone": "sl",
+  Singapore: "sg",
+  Slovakia: "sk",
+  Slovenia: "si",
+  "Solomon Islands": "sb",
+  Somalia: "so",
+  "South Africa": "za",
+  "South Korea": "kr",
+  "South Sudan": "ss",
+  Spain: "es",
+  "Sri Lanka": "lk",
+  Sudan: "sd",
+  Suriname: "sr",
+  Sweden: "se",
+  Switzerland: "ch",
+  Syria: "sy",
+  Taiwan: "tw",
+  Tajikistan: "tj",
+  Tanzania: "tz",
+  Thailand: "th",
+  "Timor-Leste": "tl",
+  Togo: "tg",
+  Tonga: "to",
+  "Trinidad and Tobago": "tt",
+  Tunisia: "tn",
+  Turkey: "tr",
+  Turkmenistan: "tm",
+  Tuvalu: "tv",
+  Uganda: "ug",
+  Ukraine: "ua",
+  "United Arab Emirates": "ae",
+  "United Kingdom": "gb",
+  "United States of America": "us",
+  "United States": "us",
+  Uruguay: "uy",
+  Uzbekistan: "uz",
+  Vanuatu: "vu",
+  "Vatican City": "va",
+  Venezuela: "ve",
+  Vietnam: "vn",
+  Yemen: "ye",
+  Zambia: "zm",
+  Zimbabwe: "zw",
+};
+
+
+export function getCountryCode(countryName?: string | null): string | null {
+  if (!countryName) return null;
+
+  const exactMatch = COUNTRY_CODE_MAP[countryName];
+  if (exactMatch) return exactMatch;
+
+  const lowerCountry = countryName.toLowerCase();
+  for (const [key, value] of Object.entries(COUNTRY_CODE_MAP)) {
+    if (key.toLowerCase() === lowerCountry) {
+      return value;
+    }
+  }
+
+  return countryName
+    .toLowerCase()
+    .replace(/[^a-z\s]/g, "")
+    .replace(/\s+/g, "-")
+    .substring(0, 2);
+}
+
