+++ app/src/lib/logger.ts
@@ -1,0 +1,15 @@
+export const logger = {
+  info: (...args: unknown[]) => {
+    if (process.env.NODE_ENV !== "production") {
+      console.log(...args);
+    }
+  },
+  warn: (...args: unknown[]) => {
+    if (process.env.NODE_ENV !== "production") {
+      console.warn(...args);
+    }
+  },
+  error: (...args: unknown[]) => {
+    console.error(...args);
+  },
+};
